// src/models/files/soundfont2/metadata.json.d.ts
interface Metadata {
  id: ['phdr', 'pbag', 'pmod', 'pgen', 'inst', 'ibag', 'imod', 'igen', 'shdr', 'smpl'],
  metadataId: ['ifil', 'isng', 'INAM', 'ICRD', 'IENG', 'IPRD', 'ICOP', 'ICMT', 'ISFT'],
  bag: ['ibag', 'pbag'],
  mod: ['imod', 'pmod'],
  gen: ['igen', 'pgen'],
  generators: [
    {name: 'startAddrsOffset', default: 0, uint: 'smpls'},
    {name: 'endAddrsOffset', default: 0, uint: 'smpls'},
    {name: 'startloopAddrsOffset', default: 0, uint: 'smpls'},
    {name: 'endloopAddrsOffset', default: 0, uint: 'smpls'},
    {name: 'startAddrsCoarseOffset', default: 0, uint: '32k smpls'},
    {name: 'modLfoToPitch', default: 0, uint: 'cent fs'},
    {name: 'vibLfoToPitch', default: 0, uint: 'cent fs'},
    {name: 'modEnvToPitch', default: 0, uint: 'cent fs'},
    {name: 'initialFilterFc', default: 13500, uint: 'cent'},
    {name: 'initialFilterQ', default: 0, uint: 'cB'},
    {name: 'modLfoToFilterFc', default: 0, uint: 'cent fs'},
    {name: 'modEnvToFilterFc', default: 0, uint: 'cent fs'},
    {name: 'endAddrsCoarseOffset', default: 0, uint: '32k smpls'},
    {name: 'modLfoToVolume', default: 0, uint: 'cB fs'},
    null,
    {name: 'chorusEffectsSend', default: 0, uint: 0.1},
    {name: 'reverbEffectsSend', default: 0, uint: 0.1},
    {name: 'pan', default: 0, uint: 0.1},
    null,
    null,
    null,
    {name: 'delayModLFO', default: -12000, uint: 'timecent'},
    {name: 'freqModLFO', default: 0, uint: 'cent'},
    {name: 'delayVibLFO', default: -12000, uint: 'timecent'},
    {name: 'freqVibLFO', default: 0, uint: 'cent'},
    {name: 'delayModEnv', default: -12000, uint: 'timecent'},
    {name: 'attackModEnv', default: -12000, uint: 'timecent'},
    {name: 'holdModEnv', default: -12000, uint: 'timecent'},
    {name: 'decayModEnv', default: -12000, uint: 'timecent'},
    {name: 'sustainModEnv', default: 0, uint: -0.1},
    {name: 'releaseModEnv', default: -12000, uint: 'timecent'},
    {name: 'keynumToModEnvHold', default: 0, uint: 'tcent/key'},
    {name: 'keynumToModEnvDecay', default: 0, uint: 'tcent/key'},
    {name: 'delayVolEnv', default: -12000, uint: 'timecent'},
    {name: 'attackVolEnv', default: -12000, uint: 'timecent'},
    {name: 'holdVolEnv', default: -12000, uint: 'timecent'},
    {name: 'decayVolEnv', default: -12000, uint: 'timecent'},
    {name: 'sustainVolEnv', default: 0, uint: 'cB attan'},
    {name: 'releaseVolEnv', default: -12000, uint: 'timecent'},
    {name: 'keynumToVolEnvHold', default: 0, uint: 'tcent/key'},
    {name: 'keynumToVolEnvDecay', default: 0, uint: 'tcent/key'},
    {name: 'instrument'},
    null,
    {name: 'keyRange', default: {lo: 0, hi: 127}, uint: 'MIDI ky#'},
    {name: 'velRange', default: {lo: 0, hi: 127}, uint: 'MIDI vel'},
    {name: 'startloopAddrsCoarseOffset', default: 0, uint: 'smpls'},
    {name: 'keynum', default: -1, uint: 'MIDI ky#'},
    {name: 'velocity', default: -1, uint: 'MIDI vel'},
    {name: 'initialAttenuation', default: 0, uint: 'cB'},
    null,
    {name: 'endloopAddrsCoarseOffset', default: 0, uint: 'smpls'},
    {name: 'coarseTune', default: 0, uint: 'semitone'},
    {name: 'fineTune', default: 0, uint: 'cent'},
    {name: 'sampleID'},
    {name: 'sampleModes', default: 0, uint: 'BitFlags'},
    null,
    {name: 'scaleTuning', default: 100, uint: 'cent/key'},
    {name: 'exclusiveClass', default: 0, uint: 'arbitary#'},
    {name: 'overridingRootKey', default: -1, uint: 'MIDI ky#'},
    null,
    {name: 'endOver'}
  ],
  sourceTypes: ['Linear', 'Concave', 'Convex', 'Switch'],
  modulatorTransforms: ['Linear', null, 'Absolute'],
  sourceEnumeratorControllerPalettes: [
    'No Controller',
    null,
    'Note-On Velocity',
    'Note-On Key Number',
    null,
    null,
    null,
    null,
    null,
    null,
    'Poly Pressure',
    null,
    null,
    'Channel Pressure',
    'Pitch Wheel',
    null,
    'Pitch Wheel Sensitivity',
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    'Link'
  ]
}

declare const Metadata: Metadata;